+++ dashboard-spring/src/main/java/com/humintecTest/dashboard/vo/StorageFreeVo.java
@@ -2,16 +2,15 @@ package com.humintecTest.dashboard.vo;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.sql.Date;
 
 @Data
 public class StorageFreeVo {
     private Date date;
     private String storage_name;
-    private BigInteger daily_used;
-    private float free;
-    private BigInteger total;
+    private long daily_used;
+    private double free;
+    private long total;
     private String start_date;
     private String end_date;
 }
